Commits not commited commited

+++ Math/Arithmetic/Mul.java
@@ -19,9 +19,9 @@ public class Mul extends Operation{
     public Construct derive() {
         var fox = this.operands.first;
         var gox = this.operands.second;
-        if (fox.derive().eval().equals(0) && gox.derive().eval().equals(0)) return new Constant();
-        else if (fox.derive().eval().equals(0)) return new Mul(gox.derive(), fox);
-        else if (gox.derive().eval().equals(0)) return new Mul(fox.derive(), gox);
+        if (fox.derive().eval().equals(BigDecimal.valueOf(0)) && gox.derive().eval().equals(BigDecimal.valueOf(0))) return new Constant();
+        else if (fox.derive().eval().equals(BigDecimal.valueOf(0))) return new Mul(gox.derive(), fox);
+        else if (gox.derive().eval().equals(BigDecimal.valueOf(0))) return new Mul(fox.derive(), gox);
         return new Add( new Mul(fox.derive(), gox), new Mul(gox.derive(), fox));
     }   
 }
